pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "meta-ads-mcp"
version = "0.1.0"
description = "MCP server for Meta Ads Manager — upload media and create campaigns via Claude"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.3.0",
    "facebook-business>=20.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
meta-ads-mcp = "meta_ads_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/meta_ads_mcp"]
